// src/api/generated/core/models/ConfigurationDetailedDto.ts
export interface ConfigurationDetailedDto {
    /**
     * 
     * @type {number}
     * @memberof ConfigurationDetailedDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigurationDetailedDto
     */
    namespaceId: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigurationDetailedDto
     */
    applicationId: number;
    /**
     * 
     * @type {string}
     * @memberof ConfigurationDetailedDto
     */
    name: string;
    /**
     * 
     * @type {SourceType}
     * @memberof ConfigurationDetailedDto
     */
    schemaSourceType: SourceType;
    /**
     * 
     * @type {string}
     * @memberof ConfigurationDetailedDto
     */
    commitVersion?: string;
    /**
     * 
     * @type {number}
     * @memberof ConfigurationDetailedDto
     */
    commitId?: number;
    /**
     * Configuration json
     * @type {string}
     * @memberof ConfigurationDetailedDto
     */
    jsonValues?: string;
    /**
     * Configuration json schema
     * @type {string}
     * @memberof ConfigurationDetailedDto
     */
    jsonSchema?: string;
}