// src/api/generated/core/models/NamespaceDto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * RTCMS4J Core-service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NamespaceDto
 */
export interface NamespaceDto {
    /**
     * 
     * @type {number}
     * @memberof NamespaceDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof NamespaceDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof NamespaceDto
     */
    description: string;
}

/**
 * Check if a given object implements the NamespaceDto interface.
 */
export function instanceOfNamespaceDto(value: object): value is NamespaceDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    return true;
}

export function NamespaceDtoFromJSON(json: any): NamespaceDto {
    return NamespaceDtoFromJSONTyped(json, false);
}

export function NamespaceDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NamespaceDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'],
    };
}

export function NamespaceDtoToJSON(json: any): NamespaceDto {
    return NamespaceDtoToJSONTyped(json, false);
}

export function NamespaceDtoToJSONTyped(value?: NamespaceDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
    };
}